Snapshot 92.0.4515.40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2929081
Bot-Commit: Chrome Release Bot (LUCI) <chrome-official-brancher@chops-service-accounts.iam.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/4515@{#177}
Cr-Branched-From: 488fc70865ddaa05324ac00a54a6eb783b4bc41c-refs/heads/master@{#885287}

+++ src/chrome/browser/ash/crosapi/browser_data_migrator.cc
@@ -8,41 +8,82 @@
 #include <utility>
 
 #include "base/containers/contains.h"
+#include "base/files/file.h"
 #include "base/files/file_enumerator.h"
 #include "base/files/file_util.h"
+#include "base/metrics/histogram_macros.h"
 #include "base/path_service.h"
+#include "base/strings/string_piece.h"
 #include "base/strings/stringprintf.h"
 #include "base/system/sys_info.h"
 #include "base/task/post_task.h"
 #include "base/task/thread_pool.h"
-#include "base/timer/elapsed_timer.h"
+#include "base/threading/thread_restrictions.h"
 #include "chrome/browser/ash/crosapi/browser_util.h"
+#include "chrome/browser/ash/profiles/profile_helper.h"
+#include "chrome/browser/browser_process.h"
 #include "chrome/common/chrome_paths.h"
+#include "components/account_id/account_id.h"
 #include "components/user_manager/user_manager.h"
+#include "components/version_info/version_info.h"
 
 namespace ash {
 namespace {
-// The name of temporary directory that will store copies of files from user
-// data directory. At the end of the migration, it will be moved to the
-// appropriate destination.
+// The name of temporary directory that will store copies of files from the
+// original user data directory. At the end of the migration, it will be moved
+// to the appropriate destination.
 constexpr char kTmpDir[] = "browser_data_migrator";
-// The base names of files and directories directly under the original user data
-// directory that should not be copied. e.g. caches or files only needed by ash.
+// The base names of files and directories directly under the original profile
+// data directory that should not be copied. e.g. caches or files only needed by
+// ash.
 const char* const kNoCopyPaths[] = {kTmpDir, "Downloads", "Cache"};
+// The base names of files and directories directory under the user data
+// directory.
+const char* const kCopyUserDataPaths[] = {"First Run"};
+
+// Lacros' user data is backward compatible up until this version.
+constexpr char kRequiredDataVersion[] = "92.0.0.0";
+
+// Copies `item` to location pointed by `dest`. Returns true on success and
+// false on failure.
+bool CopyTargetItem(const BrowserDataMigrator::TargetItem& item,
+                    const base::FilePath& dest) {
+  if (item.is_directory) {
+    if (base::CopyDirectory(item.path, dest, true /* recursive */))
+      return true;
+  } else {
+    if (base::CopyFile(item.path, dest))
+      return true;
+  }
+
+  PLOG(ERROR) << "Copy failed for " << item.path;
+  return false;
+}
 }  // namespace
+
+BrowserDataMigrator::TargetItem::TargetItem(base::FilePath path,
+                                            ItemType item_type)
+    : path(path), is_directory(item_type == ItemType::kDirectory) {}
+
+bool BrowserDataMigrator::TargetItem::operator==(const TargetItem& rhs) const {
+  return this->path == rhs.path && this->is_directory == rhs.is_directory;
+}
 
 BrowserDataMigrator::TargetInfo::TargetInfo() : total_byte_count(0) {}
 BrowserDataMigrator::TargetInfo::TargetInfo(const TargetInfo&) = default;
 BrowserDataMigrator::TargetInfo::~TargetInfo() = default;
 
 // static
-void BrowserDataMigrator::MaybeMigrate(const UserContext& user_context,
+void BrowserDataMigrator::MaybeMigrate(const AccountId& account_id,
+                                       const std::string& user_id_hash,
+                                       bool async,
                                        base::OnceClosure callback) {
   // Get the current user.
   const user_manager::User* user =
-      user_manager::UserManager::Get()->FindUser(user_context.GetAccountId());
+      user_manager::UserManager::Get()->FindUser(account_id);
   if (!user || !IsMigrationRequiredOnUI(user)) {
     // If lacros isn't enabled, skip migration and move on to the next step.
+    RecordStatus(FinalStatus::kSkipped);
     std::move(callback).Run();
     return;
   }
@@ -50,22 +91,46 @@ void BrowserDataMigrator::MaybeMigrate(const UserContext& user_context,
   base::FilePath user_data_dir;
   if (!base::PathService::Get(chrome::DIR_USER_DATA, &user_data_dir)) {
     LOG(ERROR) << "Could not get the original UDD path. Aborting migration.";
+    RecordStatus(FinalStatus::kGetPathFailed);
     std::move(callback).Run();
     return;
   }
 
-  base::FilePath profile_data_dir = user_data_dir.Append(
-      base::StringPrintf("u-%s", user_context.GetUserIDHash().c_str()));
+  // Use `GetUserProfileDir()` to manually get base name for profile dir since
+  // `MaybeMigrate()` is called before profile creation.
+  base::FilePath profile_data_dir =
+      user_data_dir.Append(ProfileHelper::GetUserProfileDir(user_id_hash));
 
   std::unique_ptr<BrowserDataMigrator> browser_data_migrator =
       std::make_unique<BrowserDataMigrator>(profile_data_dir);
 
-  base::ThreadPool::PostTaskAndReplyWithResult(
-      FROM_HERE, {base::MayBlock()},
-      base::BindOnce(&BrowserDataMigrator::MigrateInternal,
-                     std::move(browser_data_migrator)),
-      base::BindOnce(&BrowserDataMigrator::MigrateInternalFinishedUIThread,
-                     std::move(callback)));
+  // Check if user data directory needs to be wiped for a backward incompatible
+  // update.
+  base::Version data_version = crosapi::browser_util::GetDataVer(
+      g_browser_process->local_state(), user_id_hash);
+  base::Version current_version = version_info::GetVersion();
+  base::Version required_version =
+      base::Version(base::StringPiece(kRequiredDataVersion));
+  bool is_data_wipe_required =
+      IsDataWipeRequired(data_version, current_version, required_version);
+
+  if (async) {
+    base::ThreadPool::PostTaskAndReplyWithResult(
+        FROM_HERE,
+        {base::MayBlock(), base::TaskPriority::USER_VISIBLE,
+         base::TaskShutdownBehavior::BLOCK_SHUTDOWN},
+        base::BindOnce(&BrowserDataMigrator::MigrateInternal,
+                       std::move(browser_data_migrator), is_data_wipe_required),
+        base::BindOnce(&BrowserDataMigrator::MigrateInternalFinishedUIThread,
+                       std::move(callback), user_id_hash));
+  } else {
+    // Temporarily allowing blocking since we have to ensure that the migration
+    // happens before profile is created.
+    base::ScopedAllowBlocking allow_blocking;
+    MigrationResult result =
+        browser_data_migrator->MigrateInternal(is_data_wipe_required);
+    MigrateInternalFinishedUIThread(std::move(callback), user_id_hash, result);
+  }
 }
 
 // static
@@ -76,20 +141,78 @@ bool BrowserDataMigrator::IsMigrationRequiredOnUI(
 
 BrowserDataMigrator::BrowserDataMigrator(const base::FilePath& from)
     : from_dir_(from),
-      to_dir_(from.Append(kLacrosProfileDir)),
+      to_dir_(from.Append(kLacrosDir)),
       tmp_dir_(from.Append(kTmpDir)) {}
 
 BrowserDataMigrator::~BrowserDataMigrator() = default;
 
-// TODO(crbug.com/1178702): Once testing phase is over and lacros become the
+// static
+void BrowserDataMigrator::RecordStatus(const FinalStatus& final_status,
+                                       const TargetInfo* target_info,
+                                       const base::ElapsedTimer* timer) {
+  // Record final status enum.
+  UMA_HISTOGRAM_ENUMERATION(kFinalStatus, final_status);
+
+  if (!target_info)
+    return;
+  // Record byte size. Range 0 ~ 10GB in MBs.
+  UMA_HISTOGRAM_CUSTOM_COUNTS(kCopiedDataSize,
+                              target_info->total_byte_count / 1024 / 1024, 1,
+                              10000, 100);
+
+  if (!timer || final_status != FinalStatus::kSuccess)
+    return;
+  // Record elapsed time only for successful cases.
+  UMA_HISTOGRAM_MEDIUM_TIMES(kTotalTime, timer->Elapsed());
+}
+
+bool BrowserDataMigrator::IsDataWipeRequired(
+    base::Version data_version,
+    const base::Version& current_version,
+    const base::Version& required_version) {
+  // `data_version` is invalid if any wipe has not been recorded yet. In
+  // such a case, assume that the last data wipe happened significantly long
+  // time ago.
+  if (!data_version.IsValid()) {
+    data_version = base::Version("0");
+  }
+
+  if (current_version < required_version) {
+    // If `current_version` is smaller than the `required_version`, that means
+    // that the data wipe doesn't need to happen yet.
+    return false;
+  }
+
+  if (data_version >= required_version) {
+    // If `data_version` is greater or equal to `required_version`, this means
+    // data wipe has already happened and that user data is compatible with the
+    // current lacros.
+    return false;
+  }
+
+  return true;
+}
+
+// TODO(crbug.com/1178702): Once testing phase is over and lacros becomes the
 // only web browser, update the underlying logic of migration from copy to move.
 // Note that during testing phase we are copying files and leaving files in
 // original location intact. We will allow these two states to diverge.
-bool BrowserDataMigrator::MigrateInternal() {
-  if (!IsMigrationRequiredOnWorker())
-    return false;
+BrowserDataMigrator::MigrationResult BrowserDataMigrator::MigrateInternal(
+    bool is_data_wipe_required) {
+  if (is_data_wipe_required) {
+    if (!base::DeletePathRecursively(to_dir_)) {
+      RecordStatus(FinalStatus::kDataWipeFailed);
+      return {ResultValue::kFailed, ResultValue::kFailed};
+    }
+  }
 
-  base::ElapsedTimer timer;
+  ResultValue data_wipe_result =
+      is_data_wipe_required ? ResultValue::kSucceeded : ResultValue::kSkipped;
+
+  if (!IsMigrationRequiredOnWorker()) {
+    RecordStatus(FinalStatus::kSkipped);
+    return {data_wipe_result, ResultValue::kSkipped};
+  }
 
   // Check if tmp directory already exists and delete if it does.
   if (base::PathExists(tmp_dir_)) {
@@ -97,37 +220,54 @@ bool BrowserDataMigrator::MigrateInternal() {
                  << " already exists indicating migration was aborted on the"
                     "previous attempt.";
     if (!base::DeletePathRecursively(tmp_dir_)) {
-      return false;
+      PLOG(ERROR) << "Failed to delete tmp dir";
+      RecordStatus(FinalStatus::kDeleteTmpDirFailed);
+      return {data_wipe_result, ResultValue::kFailed};
     }
   }
 
   TargetInfo target_info = GetTargetInfo();
+  base::ElapsedTimer timer;
 
-  if (!HasEnoughDiskSpace(target_info))
-    return false;
+  if (!HasEnoughDiskSpace(target_info)) {
+    RecordStatus(FinalStatus::kNotEnoughSpace, &target_info);
+    return {data_wipe_result, ResultValue::kFailed};
+  }
 
   if (!CopyToTmpDir(target_info)) {
-    base::DeletePathRecursively(tmp_dir_);
-    return false;
+    if (base::PathExists(tmp_dir_)) {
+      base::DeletePathRecursively(tmp_dir_);
+    }
+    RecordStatus(FinalStatus::kCopyFailed, &target_info);
+    return {data_wipe_result, ResultValue::kFailed};
   }
 
   if (!MoveTmpToTargetDir()) {
-    base::DeletePathRecursively(tmp_dir_);
-    return false;
+    if (base::PathExists(tmp_dir_)) {
+      base::DeletePathRecursively(tmp_dir_);
+    }
+    RecordStatus(FinalStatus::kMoveFailed, &target_info);
+    return {data_wipe_result, ResultValue::kFailed};
   }
 
-  // TODO(crbug.com/1178702): Add UMA data collection here for success status,
-  // data size and elapsed time.
   LOG(WARNING) << "BrowserDataMigrator::Migrate took "
                << timer.Elapsed().InMilliseconds() << " ms and migrated "
                << target_info.total_byte_count / (1000 * 1000) << " MBs.";
-  return true;
+  RecordStatus(FinalStatus::kSuccess, &target_info, &timer);
+  return {data_wipe_result, ResultValue::kSucceeded};
 }
 
 void BrowserDataMigrator::MigrateInternalFinishedUIThread(
     base::OnceClosure callback,
-    bool did_migrate) {
-  if (did_migrate) {
+    const std::string& user_id_hash,
+    MigrationResult result) {
+  if (result.data_wipe == ResultValue::kSucceeded) {
+    crosapi::browser_util::RecordDataVer(g_browser_process->local_state(),
+                                         user_id_hash,
+                                         version_info::GetVersion());
+  }
+
+  if (result.data_migration == ResultValue::kSucceeded) {
     // If we did a migration, then we should set kClearUserDataDir1Pref. Note
     // that if we did the migration, then the new user-data-dir has the ash
     // profile as the main lacros profile.
@@ -158,18 +298,30 @@ BrowserDataMigrator::TargetInfo BrowserDataMigrator::GetTargetInfo() const {
   for (base::FilePath entry = enumerator.Next(); !entry.empty();
        entry = enumerator.Next()) {
     if (base::Contains(kNoCopyPaths, entry.BaseName().value())) {
-      // Skip if the base name is present in kNoCopyPaths.
+      // Skip if the base name is present in `kNoCopyPaths`.
       continue;
     }
 
     const base::FileEnumerator::FileInfo& info = enumerator.GetInfo();
     if (S_ISREG(info.stat().st_mode)) {
-      target_info.file_paths.emplace_back(entry);
+      target_info.profile_data_items.emplace_back(
+          TargetItem{entry, TargetItem::ItemType::kFile});
       target_info.total_byte_count += info.GetSize();
     } else {
       // Treat symlink the same as directory since even if it points to a file,
-      // both ComputeDirectorySize and CopyDirectory can be used.
-      target_info.dir_paths.emplace_back(entry);
+      // both `ComputeDirectorySize()` and `CopyDirectory()` can be used.
+      target_info.profile_data_items.emplace_back(
+          TargetItem{entry, TargetItem::ItemType::kDirectory});
+      target_info.total_byte_count += base::ComputeDirectorySize(entry);
+    }
+  }
+
+  // Copy files directly under user data directory.
+  for (auto* copy_path : kCopyUserDataPaths) {
+    base::FilePath entry = from_dir_.DirName().Append(copy_path);
+    if (base::PathExists(entry)) {
+      target_info.user_data_items.emplace_back(
+          TargetItem{entry, TargetItem::ItemType::kFile});
       target_info.total_byte_count += base::ComputeDirectorySize(entry);
     }
   }
@@ -191,24 +343,28 @@ bool BrowserDataMigrator::HasEnoughDiskSpace(
 
 bool BrowserDataMigrator::CopyToTmpDir(const TargetInfo& target_info) const {
   base::File::Error error;
-  if (!base::CreateDirectoryAndGetError(tmp_dir_, &error)) {
+  if (!base::CreateDirectoryAndGetError(tmp_dir_.Append(kLacrosProfilePath),
+                                        &error)) {
     PLOG(ERROR) << "CreateDirectoryFailed " << error;
+    // Maps to histogram enum `PlatformFileError`.
+    UMA_HISTOGRAM_ENUMERATION(kCreateDirectoryFail, -error,
+                              -base::File::FILE_ERROR_MAX);
     return false;
   }
 
-  for (const auto& target_file : target_info.file_paths) {
-    if (!base::CopyFile(target_file, tmp_dir_.Append(target_file.BaseName()))) {
-      PLOG(ERROR) << "CopyFile failed for " << target_file;
+  for (const auto& target_item : target_info.profile_data_items) {
+    base::FilePath dest =
+        tmp_dir_.Append(kLacrosProfilePath).Append(target_item.path.BaseName());
+
+    if (!CopyTargetItem(target_item, dest))
       return false;
-    }
   }
 
-  for (const auto& target_dir : target_info.dir_paths) {
-    if (!base::CopyDirectory(target_dir, tmp_dir_.Append(target_dir.BaseName()),
-                             true /* recursive */)) {
-      PLOG(ERROR) << "CopyDirectory failed for " << target_dir;
+  for (const auto& target_item : target_info.user_data_items) {
+    base::FilePath dest = tmp_dir_.Append(target_item.path.BaseName());
+
+    if (!CopyTargetItem(target_item, dest))
       return false;
-    }
   }
 
   return true;
