+++ src/chrome/browser/resources/chromeos/accessibility/chromevox/background/command_handler.js
@@ -244,14 +244,28 @@ CommandHandler.onCommand = function(command) {
     case 'help':
       (new PanelCommand(PanelCommandType.TUTORIAL)).send();
       return false;
-    case 'toggleDarkScreen':
+    case 'toggleScreen':
       const oldState = sessionStorage.getItem('darkScreen');
       const newState = (oldState === 'true') ? false : true;
-      sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
-      chrome.accessibilityPrivate.darkenScreen(newState);
-      new Output()
-          .format((newState) ? '@darken_screen' : '@undarken_screen')
-          .go();
+      if (newState && localStorage['acceptToggleScreen'] !== 'true') {
+        // If this is the first time, show a confirmation dialog.
+        chrome.accessibilityPrivate.showConfirmationDialog(
+            Msgs.getMsg('toggle_screen_title'),
+            Msgs.getMsg('toggle_screen_description'), (confirmed) => {
+              if (confirmed) {
+                sessionStorage.setItem('darkScreen', 'true');
+                localStorage['acceptToggleScreen'] = true;
+                chrome.accessibilityPrivate.darkenScreen(true);
+                new Output().format('@toggle_screen_off').go();
+              }
+            });
+      } else {
+        sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
+        chrome.accessibilityPrivate.darkenScreen(newState);
+        new Output()
+            .format((newState) ? '@toggle_screen_off' : '@toggle_screen_on')
+            .go();
+      }
       return false;
     case 'toggleSpeechOnOrOff':
       const state = ChromeVox.tts.toggleSpeechOnOrOff();
@@ -707,7 +721,7 @@ CommandHandler.onCommand = function(command) {
                       .withoutHints()
                       .withRichSpeechAndBraille(
                           ChromeVoxState.instance.currentRange, prevRange,
-                          Output.EventType.NAVIGATE)
+                          OutputEventType.NAVIGATE)
                       .onSpeechEnd(continueReading);
 
         if (!o.hasSpeech) {
@@ -725,7 +739,7 @@ CommandHandler.onCommand = function(command) {
             new Output()
                 .withoutHints()
                 .withRichSpeechAndBraille(
-                    collapsedRange, collapsedRange, Output.EventType.NAVIGATE)
+                    collapsedRange, collapsedRange, OutputEventType.NAVIGATE)
                 .onSpeechEnd(continueReading);
 
         if (o.hasSpeech) {
@@ -820,7 +834,7 @@ CommandHandler.onCommand = function(command) {
           const o =
               new Output()
                   .format('@end_selection')
-                  .withSpeechAndBraille(sel, sel, Output.EventType.NAVIGATE)
+                  .withSpeechAndBraille(sel, sel, OutputEventType.NAVIGATE)
                   .go();
           DesktopAutomationHandler.instance.ignoreDocumentSelectionFromAction(
               false);
@@ -832,7 +846,7 @@ CommandHandler.onCommand = function(command) {
     case 'fullyDescribe':
       const o = new Output();
       o.withContextFirst()
-          .withRichSpeechAndBraille(current, null, Output.EventType.NAVIGATE)
+          .withRichSpeechAndBraille(current, null, OutputEventType.NAVIGATE)
           .go();
       return false;
     case 'viewGraphicAsBraille':
@@ -1037,24 +1051,13 @@ CommandHandler.onCommand = function(command) {
         }
       }
 
-      // Get unicode-aware array of characters.
-      const characterArray = [...word];
       const language = chrome.i18n.getUILanguage();
-      for (let i = 0; i < characterArray.length; ++i) {
-        const character = characterArray[i];
-        const phoneticText = PhoneticData.forCharacter(character, language);
-        // Speak the character followed by its phonetic disambiguation, if it
-        // was found.
+      const phoneticText = PhoneticData.forText(word, language);
+      if (phoneticText) {
         new Output()
-            .withString(character)
-            .withQueueMode(i === 0 ? QueueMode.CATEGORY_FLUSH : QueueMode.QUEUE)
+            .withString(phoneticText)
+            .withQueueMode(QueueMode.CATEGORY_FLUSH)
             .go();
-        if (phoneticText) {
-          new Output()
-              .withString(phoneticText)
-              .withQueueMode(QueueMode.QUEUE)
-              .go();
-        }
       }
     }
       return false;
